generalize add-to-cart function
