+++ HW_Report/shopping_site/react/src/App.js
@@ -7,7 +7,7 @@ import ProductDetail from "./pages/ProductDetail";
 import ContactPage from "./pages/ContactPage";
 import CartPage from "./pages/CartPage";
 import AboutPage from "./pages/AboutPage";
-import { ToastContainer } from "react-toastify";
+import { ToastContainer, toast, Bounce } from "react-toastify";
 
 function App() {
   const [products, setProducts] = useState([]);
@@ -25,6 +25,17 @@ function App() {
       .then((res) => setProducts(res.data))
       .catch((err) => console.log("Error fetching products:", err));
   }, []);
+
+  const notify = (messege, type = "info") => {
+    toast[type](messege, {
+      position: "bottom-right",
+      autoClose: 3000,
+      draggable: false,
+      hideProgressBar: true,
+      closeOnClick: true,
+      transition: Bounce,
+    });
+  };
 
   const checkToggle = (product, value) => {
     setCart((c) => {
@@ -56,17 +67,35 @@ function App() {
     setCart((c) => c.filter((p) => p.id !== product.id));
   };
 
+  const handleAddToCart = (product, selectedQuantity = 1) => {
+    let amount = cart.find((p) => p.id === product.id)?.amount;
+    if (amount && selectedQuantity + amount > product.stock) {
+      for (let i = 0; i < product.stock - amount; i++) addCart(product);
+      notify(
+        `您選的數量超過庫存，為您加入 ${product.stock - amount} 件 ${product.name} 到購物車`,
+        "warn",
+      );
+      return;
+    }
+    for (let i = 0; i < selectedQuantity; i++) addCart(product);
+    notify(`成功加入 ${selectedQuantity} 件 ${product.name} 到購物車`);
+  };
+
   return (
     <Router>
       <Routes>
         <Route path="/" element={<Layout />}>
           <Route
             index
-            element={<HomePage products={products} addCart={addCart} />}
+            element={
+              <HomePage products={products} handleAddToCart={handleAddToCart} />
+            }
           />
           <Route
             path="product/:id"
-            element={<ProductDetail cart={cart} addCart={addCart} />}
+            element={
+              <ProductDetail cart={cart} handleAddToCart={handleAddToCart} />
+            }
           />
           <Route path="contact" element={<ContactPage />} />
           <Route path="about" element={<AboutPage />} />
