+++ HW_Report/shopping_site/react/src/components/Products.jsx
@@ -1,9 +1,8 @@
 import React from "react";
-import { toast, Bounce } from "react-toastify";
 import { Link } from "react-router-dom";
 import styles from "./Products.module.css";
 
-function Products({ products, addCart }) {
+function Products({ products, handleAddToCart }) {
   return (
     <section className={styles.products}>
       {products.length > 0 ? (
@@ -14,21 +13,7 @@ function Products({ products, addCart }) {
               <h2>{product.name}</h2>
               <h2 className={styles.price}>${product.price}</h2>
             </Link>
-            <button
-              onClick={() => {
-                addCart(product);
-                toast("成功加入購物車", {
-                  position: "bottom-right",
-                  autoClose: 5000,
-                  draggable: false,
-                  hideProgressBar: true,
-                  closeOnClick: true,
-                  transition: Bounce,
-                });
-              }}
-            >
-              加入購物車
-            </button>
+            <button onClick={() => handleAddToCart(product)}>加入購物車</button>
           </div>
         ))
       ) : (
